Factor out extensions and use in PrettyConfig to configure implicit_some
