+++ src/lib.rs
@@ -60,6 +60,7 @@ Serializing / Deserializing is as simple as calling `to_string` / `from_str`.
 pub mod de;
 pub mod ser;
 pub mod value;
+pub mod extensions;
 pub use crate::value::Value;
 
 mod parse;
