+++ src/ser/mod.rs
@@ -4,6 +4,8 @@ use std::{
     fmt::{Display, Formatter, Result as FmtResult, Write},
 };
 
+use crate::extensions::Extensions;
+
 mod value;
 
 /// Serializes `value` and returns it as string.
@@ -96,8 +98,8 @@ pub struct PrettyConfig {
     /// Enumerate array items in comments
     #[serde(default = "default_enumerate_arrays")]
     pub enumerate_arrays: bool,
-    /// Enable implicit_some extension
-    pub implicit_some: bool,
+    /// Enable extensions. Only configures 'implicit_some' for now.
+    pub extensions: Extensions,
     /// Private field to ensure adding a field is non-breaking.
     #[serde(skip)]
     _future_proof: (),
@@ -159,11 +161,11 @@ impl PrettyConfig {
         self
     }
 
-    /// Configures whether the implicit_some extension will be enabled
+    /// Configures extensions
     ///
-    /// Default: `false`
-    pub fn with_implicit_some(mut self, implicit_some: bool) -> Self {
-        self.implicit_some = implicit_some;
+    /// Default: Extensions::empty()
+    pub fn with_extensions(mut self, extensions: Extensions) -> Self {
+        self.extensions = extensions;
 
         self
     }
@@ -194,8 +196,8 @@ fn default_enumerate_arrays() -> bool {
     false
 }
 
-fn default_implicit_some() -> bool {
-    false
+fn default_extensions() -> Extensions {
+    Extensions::empty()
 }
 
 impl Default for PrettyConfig {
@@ -206,7 +208,7 @@ impl Default for PrettyConfig {
             indentor: default_indentor(),
             separate_tuple_members: default_separate_tuple_members(),
             enumerate_arrays: default_enumerate_arrays(),
-            implicit_some: default_implicit_some(),
+            extensions: default_extensions(),
             _future_proof: (),
         }
     }
@@ -229,7 +231,7 @@ impl Serializer {
     /// Most of the time you can just use `to_string` or `to_string_pretty`.
     pub fn new(config: Option<PrettyConfig>, struct_names: bool) -> Self {
         let initial_output = if let Some(conf) = &config {
-            if conf.implicit_some {
+            if conf.extensions.contains(Extensions::IMPLICIT_SOME) {
                 "#![enable(implicit_some)]".to_string() + &conf.new_line
             } else {
                 String::new()
@@ -271,10 +273,10 @@ impl Serializer {
             .map_or(false, |&(ref config, _)| config.separate_tuple_members)
     }
 
-    fn implicit_some(&self) -> bool {
+    fn extensions(&self) -> Extensions {
         self.pretty
             .as_ref()
-            .map_or(false, |&(ref config, _)| config.implicit_some)
+            .map_or(Extensions::empty(), |&(ref config, _)| config.extensions)
     }
 
     fn start_indent(&mut self) {
@@ -414,11 +416,12 @@ impl<'a> ser::Serializer for &'a mut Serializer {
     where
         T: ?Sized + Serialize,
     {
-        if !self.implicit_some() {
+        let implicit_some = self.extensions().contains(Extensions::IMPLICIT_SOME);
+        if !implicit_some {
             self.output += "Some(";
         }
         value.serialize(&mut *self)?;
-        if !self.implicit_some() {
+        if !implicit_some {
             self.output += ")";
         }
